Create badRespFinder.gs
This is a google sheets app script that goes through a sheet (built for an online spreadsheet storing the results of a series of Spanish tests taken by UMich students), finds the location of cells that are filled with a single question mark, and prints the locations along with the name and section number of the test taker on a separate sheet.  This allows the data enterer to quickly locate the questions that could not be read from the scanned PDF and need to be read from the hard paper copies of the tests.